requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy<2.0
pymysql>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.5
pandas>=1.3.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
bcrypt>=3.2.0
email-validator>=1.1.3
requests>=2.25.0